优化部分单元测试代码，以规避潜在的 bug
com.dwarfeng.fdr.impl.service.FilterInfoMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.FilterSupportMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.MapperSupportMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.PointMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.TriggerInfoMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.TriggerSupportMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.WasherInfoMaintainServiceImplTest
com.dwarfeng.fdr.impl.service.WasherSupportMaintainServiceImplTest

+++ fdr-impl/src/test/java/com/dwarfeng/fdr/impl/service/FilterInfoMaintainServiceImplTest.java
@@ -15,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -59,9 +60,14 @@ public class FilterInfoMaintainServiceImplTest {
             }
         } finally {
             for (FilterInfo filterInfo : filterInfos) {
+                if (Objects.isNull(filterInfo.getKey())) {
+                    continue;
+                }
                 filterInfoMaintainService.deleteIfExists(filterInfo.getKey());
             }
-            pointMaintainService.deleteIfExists(parentPoint.getKey());
+            if (Objects.nonNull(parentPoint.getKey())) {
+                pointMaintainService.deleteIfExists(parentPoint.getKey());
+            }
         }
     }
 }
